requirements.txt
annotated-types==0.7.0
anyio==4.10.0
assemblyai==0.43.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
defusedxml==0.7.1
dnspython==2.8.0
fonttools==4.59.2
fpdf2==2.8.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
numpy==2.3.3
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
pyparsing==3.2.4
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
shiboken6==6.9.2
sniffio==1.3.1
sounddevice==0.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wavio==0.0.9
websockets==15.0.1
